tools: aggregatedStats tests working
