+++ biodata-formats/src/test/java/org/opencb/biodata/formats/annotation/io/VepFormatReaderTest.java
@@ -1,4 +1,3 @@
-
 /*
  * Copyright 2015 OpenCB
  *
